+++ src/main/kotlin/de/itcr/termite/model/repository/FhirCodeSystemMetadataRepository.kt
@@ -3,4 +3,4 @@ package de.itcr.termite.model.repository
 import de.itcr.termite.model.entity.FhirCodeSystemMetadata
 import org.springframework.data.repository.CrudRepository
 
-interface FhirCodeSystemMetadataRepository: CrudRepository<FhirCodeSystemMetadata, Long>
+interface FhirCodeSystemMetadataRepository: CrudRepository<FhirCodeSystemMetadata, Int>
